layers/ai/market-sentiment-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
vaderSentiment==3.3.2
google-generativeai==0.8.4
